Cleaned up support for durations in time units

+++ support/cas-server-support-wsfederation/src/main/java/org/apereo/cas/support/wsfederation/WsFederationConfiguration.java
@@ -63,7 +63,7 @@ public class WsFederationConfiguration implements Serializable {
     
     private String relyingPartyIdentifier;
     
-    private int tolerance;
+    private long tolerance;
     
     private WsFedPrincipalResolutionAttributesType attributesType;
     
@@ -185,7 +185,7 @@ public class WsFederationConfiguration implements Serializable {
      *
      * @return the tolerance in milliseconds
      */
-    public int getTolerance() {
+    public long getTolerance() {
         return this.tolerance;
     }
 
@@ -194,7 +194,7 @@ public class WsFederationConfiguration implements Serializable {
      *
      * @param tolerance the tolerance in milliseconds
      */
-    public void setTolerance(final int tolerance) {
+    public void setTolerance(final long tolerance) {
         this.tolerance = tolerance;
     }
